12.28-f : Full implementation of new menu system
QR code template is now configurable from the QRTEMPLATE settings
Large speedup of SPI transmissions on OLED screen (x10)

+++ firmware/display/display.cpp
@@ -62,7 +62,7 @@ void display_draw_point(int x,int y,uint16_t color) {
   write_d(color);
 }
 
-void display_draw_rectangle(int start_x,int start_y,int end_x,int end_y,uint16_t color) {
+void display_draw_rectangle( uint8_t start_x, uint8_t start_y, uint8_t end_x, uint8_t end_y,uint16_t color) {
 
   if(start_x < 0  ) start_x = 0;
   if(start_y < 0  ) start_y = 0;
@@ -74,10 +74,10 @@ void display_draw_rectangle(int start_x,int start_y,int end_x,int end_y,uint16_t
   Set_Column_Address(start_x, end_x);
   Set_Row_Address   (start_y, end_y);
 
-  uint32_t size = (end_x-start_x+1)*(end_y-start_y+1);
+  uint16_t size = (end_x-start_x+1)*(end_y-start_y+1);
 
   write_c(0x5C);    // Enable MCU to Read from RAM
-  for (uint32_t i=0; i<size;i++) {
+  for (uint16_t i = 0; i < size; i++) {
     write_d(color);
     write_d(color >> 8);
   }
@@ -252,12 +252,12 @@ void display_draw_fixedimage(uint8_t x,uint8_t y,uint8_t image_number,uint16_t b
 void display_test() {
   display_clear(0);
 
-  for(int x=0;x<128;x++) {
+  for(uint8_t x=0;x<128;x++) {
     display_draw_rectangle(x,0,x,127,0xFFFF);
     delay_us(400000);
     display_draw_rectangle(x,0,x,127,0x0000);
   }
-  for(int y=0;y<128;y++) {
+  for(uint8_t y=0;y<128;y++) {
     display_draw_rectangle(0,y,127,y,0xFFFF);
     delay_us(400000);
     display_draw_rectangle(0,y,127,y,0x0000);
