+++ firmware/display/display.h
@@ -12,7 +12,7 @@ void display_clear(int16 color);
 void display_draw_image(int x,int y,int width,int height, uint16 *image_data);
 void display_draw_line(int start_x,int start_y,int end_x,int end_y,uint16_t color=65535);
 void display_draw_point(int x,int y,uint16_t color=65535);
-void display_draw_rectangle(int start_x,int start_y,int end_x,int end_y,uint16_t color);
+void display_draw_rectangle(uint8_t start_x,uint8_t start_y,uint8_t end_x,uint8_t end_y,uint16_t color);
 void display_draw_text( int x,int y,const char *text, int16_t foreground, int16_t background);
 void display_draw_text_center(int y,const char *text, int16_t foreground, int16_t background);
 void display_draw_number(int x,int y,uint32_t number,int width,int16_t foreground, int16_t background);
